Break API.scala into parts.
